requirements.txt
Django==4.1.2
loguru==0.6.0
Pillow==9.2.0
django-cleanup==6.0.0
django-debug-toolbar==3.7.0
mysqlclient==2.1.1
openpyxl==3.1.0
django-tinymce==3.5.0
PyMuPDF==1.21.1
djangorestframework==3.14.0
django-cors-headers==4.2.0